(titus) move lookup of security group details to cluster caching time and remove instance caching agent since the functionality there is redundant.
