+++ clouddriver-titus/src/main/groovy/com/netflix/spinnaker/clouddriver/titus/model/TitusServerGroup.groovy
@@ -61,6 +61,7 @@ class TitusServerGroup implements ServerGroup, Serializable {
   TitusServerGroup(Job job, String account, String region) {
     id = job.id
     name = job.name
+    securityGroupDetails = job.securityGroupDetails
     image << [dockerImageName: job.applicationName]
     image << [dockerImageVersion: job.version]
     entryPoint = job.entryPoint
